--- src/main/java/com/starpony/prohojemba/users/Profile.java
@@ -1,23 +1,0 @@
-package com.starpony.prohojemba.users;
-
-
-public class Profile {
-    private String username;
-    private String avatar;
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getAvatar() {
-        return avatar;
-    }
-
-    public void setAvatar(String avatar) {
-        this.avatar = avatar;
-    }
-}
